Refacto utility functions used in component into their own files
